pyproject.toml
[project]
name = "influencer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "psycopg[binary]>=3.2.9",
    "requests>=2.32.5",
    "tiktokapi>=7.3.2",
]
